requirements.txt
streamlit
spacy
pandas
mysql-connector-python
langdetect
deep-translator
reportlab
PyMuPDF
transformers==4.30.2
sentence-transformers==2.2.2
tokenizers==0.13.3
sentencepiece==0.1.99
torch>=1.12.0
plotly
keybert
scikit-learn

